Fold spool copy-on-fanout into BroadcastExchange, the only exchange sharing one block across mailboxes
