+++ pinot-query-runtime/src/main/java/org/apache/pinot/query/mailbox/GrpcSendingMailbox.java
@@ -116,6 +116,9 @@ public class GrpcSendingMailbox implements SendingMailbox {
     return false;
   }
 
+  /// NOTE: [org.apache.pinot.query.runtime.operator.exchange.BlockExchange] implementations rely on this method
+  /// serializing the block synchronously on the calling thread: once it returns, the block's contents may be handed
+  /// by reference to a local receiver that mutates them.
   @Override
   public void send(MseBlock.Data data) {
     QueryThreadContext.checkTerminationAndSampleUsage(SEND_SCOPE);
