+++ pinot-query-runtime/src/main/java/org/apache/pinot/query/runtime/operator/MailboxSendOperator.java
@@ -88,9 +88,8 @@ public class MailboxSendOperator extends MultiStageOperator {
   ///
   /// 1. One inner exchange is created for each receiver stage, using the method mentioned above and keeping the
   ///    distribution type specified in the [MailboxSendNode].
-  /// 2. Then, a single outer spool broadcast exchange is created to fan out the data to all the inner exchanges.
-  ///    It copies blocks that carry mutable cells (aggregation intermediate results) so that no two receiver stages
-  ///    share them.
+  /// 2. Then, a single outer broadcast exchange is created to fan out the data to all the inner exchanges. It copies
+  ///    blocks that carry aggregation intermediate results so that no two receiver stages share them.
   ///
   /// @see BlockExchange#asSendingMailbox(String)
   private static BlockExchange getBlockExchange(OpChainExecutionContext ctx, MailboxSendNode node,
@@ -111,7 +110,8 @@ public class MailboxSendOperator extends MultiStageOperator {
     }
 
     Function<List<SendingMailbox>, Integer> statsIndexChooser = getStatsIndexChooser(ctx, node);
-    return BlockExchange.getSpoolExchange(perStageSendingMailboxes, mainSplitter, statsIndexChooser);
+    return BlockExchange.getExchange(perStageSendingMailboxes, RelDistribution.Type.BROADCAST_DISTRIBUTED,
+        List.of(), mainSplitter, statsIndexChooser, node.getHashFunction());
   }
 
   private static Function<List<SendingMailbox>, Integer> getStatsIndexChooser(OpChainExecutionContext ctx,
